requirements.txt
altgraph==0.17.4
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
buildozer==1.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
docutils==0.21.2
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
lxml==5.3.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
numpy==2.1.0
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pefile==2023.2.7
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
psutil==6.0.0
ptyprocess==0.7.0
PyAutoGUI==0.9.54
pycparser==2.22
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.1
pytube==15.0.0
pytweening==1.2.0
pywin32==306
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.23.1
setuptools==72.2.0
sh==2.0.7
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
youtube-dl==2021.12.17
